Agregar funcionalidad para gestionar jornadas: crear, obtener, actualizar y eliminar jornadas

+++ backend/queries/attendance.sql
@@ -1,1 +1,1 @@
-SELECT insertar_marca($1);
+SELECT insertar_marca($1::integer);
